backend/requirements.txt
async-timeout==4.0.2
bcrypt==3.2.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
Flask==2.2.3
flask-cors==3.0.10
importlib-metadata==4.6.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==8.10.0
mysql-connector-python==8.0.32
protobuf==3.12.4
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2022.1
PyYAML==6.0
requests==2.25.1
six==1.16.0
urllib3==1.26.5
Werkzeug==2.2.3
uWSGI==2.0.21